fix(profile): version is uint32 to match the new routing protocol
The profile-version field on UserProfile / UserEntry / UpdateProfileResponse
was originally typed as uint64 / u64. The new routing protocol uses
uint32 for the per-user version field on the wire — keeping these as
uint64 would either misalign on the wire (over-sized field on a
fixed-shape header) or force a lossy convert at every hop.

Switched to uint32 / u32 end-to-end:

  - protobuf/proto_definitions/router/router_net_info.proto:
    UserProfile.version uint64 -> uint32. Comment now flags the
    MUST-be-uint32 contract.
  - protobuf/proto_definitions/router/users.proto:
    UserEntry.profile_version uint64 -> uint32, comment ties it to
    UserProfile.version.
  - protobuf/proto_definitions/node/user_accounts.proto:
    UpdateProfileResponse.new_version uint64 -> uint32.
  - libqaul: User.version and UserData.version u64 -> u32 with a
    doc-comment naming the routing-protocol contract.
  - libqaul: send_update_profile_response's new_version arg u64 -> u32.
  - qauld-ctl: print_extended_profile's profile_version arg u64 -> u32.
  - Re-generated protobuf rust files match the new types.

updated_at stays u64 (ms timestamps need it).

Verified: cargo build -p libqaul -p qauld -p qauld-ctl clean;
sign_verify_profile_roundtrip / tampered_profile_rejected /
wrong_key_profile_rejected all pass.

+++ rust/clients/qauld-ctl/src/commands/users.rs
@@ -355,7 +355,7 @@ impl RpcCommand for UsersSubcmd {
 fn print_extended_profile(
     bio: &str,
     avatar: &[u8],
-    profile_version: u64,
+    profile_version: u32,
     profile_updated_at: u64,
     preferred_custody_route: &[Vec<u8>],
 ) {
